backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==5.0.1
django-cors-headers==4.3.1
django-oauth-toolkit==2.3.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-oauth==1.1.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.5.1
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
simplejson==3.19.2
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0
